Use system nss again

+++ security/nss/cmd/httpserv/httpserv.c
@@ -1321,7 +1321,7 @@ main(int argc, char **argv)
         FILE *tmpfile = fopen(pidFile, "w+");
 
         if (tmpfile) {
-            fprintf(tmpfile, "%d", (int)getpid());
+            fprintf(tmpfile, "%d", getpid());
             fclose(tmpfile);
         }
     }
